fix(review): move api route from /agent_configurations/[aId]/skill_configurations to use existing /skill_configurations
