+++ front/pages/api/w/[wId]/assistant/skill_configurations/index.ts
@@ -3,14 +3,17 @@ import * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
+import { getAgentConfiguration } from "@app/lib/api/assistant/configuration/agent";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
 import { SkillConfigurationModel } from "@app/lib/models/skill";
 import { GroupResource } from "@app/lib/resources/group_resource";
+import { SkillConfigurationResource } from "@app/lib/resources/skill_configuration_resource";
 import { makeSId } from "@app/lib/resources/string_ids";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
+import { isGlobalAgentId, isString } from "@app/types";
 import type { SkillConfigurationType } from "@app/types/skill_configuration";
 
 export type PostSkillConfigurationResponseBody = {
@@ -24,6 +27,10 @@ export type PostSkillConfigurationResponseBody = {
     | "authorId"
   >;
 };
+
+export interface GetAgentSkillsResponseBody {
+  skills: SkillConfigurationType[];
+}
 
 // Request body schema for POST
 const PostSkillConfigurationRequestBodySchema = t.type({
@@ -40,12 +47,12 @@ type PostSkillConfigurationRequestBody = t.TypeOf<
 async function handler(
   req: NextApiRequest,
   res: NextApiResponse<
-    WithAPIErrorResponse<PostSkillConfigurationResponseBody>
+    | WithAPIErrorResponse<PostSkillConfigurationResponseBody>
+    | WithAPIErrorResponse<GetAgentSkillsResponseBody>
   >,
   auth: Authenticator
 ): Promise<void> {
   const owner = auth.getNonNullableWorkspace();
-  const user = auth.getNonNullableUser();
 
   const featureFlags = await getFeatureFlags(owner);
   if (!featureFlags.includes("skills")) {
@@ -59,7 +66,52 @@ async function handler(
   }
 
   switch (req.method) {
+    case "GET": {
+      const { aId } = req.query;
+      if (!isString(aId)) {
+        return apiError(req, res, {
+          status_code: 400,
+          api_error: {
+            type: "invalid_request_error",
+            message: "Invalid agent configuration ID.",
+          },
+        });
+      }
+
+      const agent = await getAgentConfiguration(auth, {
+        agentId: aId,
+        variant: "light",
+      });
+      if (!agent) {
+        return apiError(req, res, {
+          status_code: 404,
+          api_error: {
+            type: "agent_configuration_not_found",
+            message: "The agent configuration was not found.",
+          },
+        });
+      }
+
+      if (isGlobalAgentId(agent.sId)) {
+        // TODO(skills): Implement fetching skills for global agents.
+        return res.status(200).json({
+          skills: [],
+        });
+      }
+
+      const skills =
+        await SkillConfigurationResource.fetchByAgentConfigurationId(
+          auth,
+          agent.id
+        );
+
+      return res.status(200).json({
+        skills: skills.map((s) => s.toJSON()),
+      });
+    }
     case "POST": {
+      const user = auth.getNonNullableUser();
+
       const bodyValidation = PostSkillConfigurationRequestBodySchema.decode(
         req.body
       );
@@ -133,7 +185,8 @@ async function handler(
         status_code: 405,
         api_error: {
           type: "method_not_supported_error",
-          message: "The method passed is not supported, POST is expected.",
+          message:
+            "The method passed is not supported, GET or POST is expected.",
         },
       });
   }
